+++ @types/artist.d.ts
@@ -1,0 +1,4 @@
+export type artistType = {
+  name: string;
+  profile_photo: string;
+};
