+++ @types/genre.d.ts
@@ -1,0 +1,4 @@
+export type genreType = {
+  name: string;
+  path: string;
+};
